Allow Custom Size Buffers
To be honest, I mixed up refactoring with the new feature. The new
feature allows for custom size buffers to support custom protocols which
only have to be based on size prefixed frames. I also introduced the
possibility to turn off heartbeats to support protocols without them.

Custom Size Buffers

There are new options to connect:

- :size-buf-alloc-fn :: fn allocating the size buf
- :read-size-fn      :: fn reading the size from the size-buf

The new options default to the original implementation. Personally I
need a size buffer with byte order little-endian for my protocol.
Specifying a :size-buf-alloc-fn allocating a ByteBuffer with
little-endian byte order makes this possible.

Disabling the Heartbeat

One can now disable the heartbeats by setting :heartbeat-period to nil.

Refactoring

One goal of the refactoring is to get rid of all type hints in order to
make the code more readable. I introduced the namespace net.async.nio
which provides wrapper functions for all used NIO methods. I simply
moved the type hints into these wrapper functions.

I improved the runtime of the select-opts function from about 190 ns to
about 45 ns measured which criterium. This improvement was not my
primary goal and I did not profile. Originally my goal was only to get
rid of the SelectionKey/... constants.

In the function client-socket, I replaced (go (>! ...)) with
(async/put! ...) because it does the same.

I splitted on-write-done separating the write buffer filling into the
function fill-write-bufs! because it looked awkward to have
on-write-done being called in the function new-socket.

I also preserve the size buffer for writing as it was done already for
reading. I did this because its possible and the custom size buffer
function will be called only twice, once for the read buffer und once
for the write buffer.

I added some trace logging for my own better understanding. It has
nearly no runtime overhead.

I added (set! *warn-on-reflection* true) on top of each namespace
because that is the preferred method to do this. It is not preferred to
add warn-on-reflection in project.clj.

I replaced the (ClosedChannelException.) while reading -1 which an
(IOException. "End of stream reached.") because I hit this thing once
and the missing error message was not helpful.

I added bangs to some of the side-effect functions.